// .pnpm-store/v11/projects/6ce9d4c461b402a23f966a6025831758/src/modules/league-members/league-members.types.ts
export type LeagueMember = {
    id: string;
    leagueId: string;
    userId: string;
    role: "owner" | "player" | "admin" | "spec";
    createdAt: Date;
    nickname?: string;
    gameName?: string;
    tagLine?: string;
};
export type LeagueMemberIdentity = {
    userId?: string[];
    role?: string[];
    leagueId?: string[];
    nickname?: string[];
};
export type CreateLeagueMemberDTO = {
    role: Exclude<LeagueMember['role'], 'owner'>;
};
export type UpdateLeagueMemberDTO = {
    role: LeagueMember['role'];
};